requirements.txt
pandas
openpyxl
selenium
webdriver-manager
python-dotenv